Rename to TransferBuilder. Will create a separate Transfer struct with better validation tracking complexity.
